Refactored code structure starting from folders hierarchy and finishing on skeleton of classes
