+++ ToDoList/ToDoList/Classes/SceneDelegate.swift
@@ -6,28 +6,17 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let windowScene = (scene as? UIWindowScene) else { return }
-        
         let window = UIWindow(windowScene: windowScene)
-        
         let taskViewController = TaskManagerViewController(
             nibName: String(describing: TaskManagerViewController.self),
             bundle: nil
         )
-        
         let taskNavigationController = UINavigationController(rootViewController: taskViewController)
-        
         window.rootViewController = taskNavigationController
-        
         window.makeKeyAndVisible()
-        
         self.window = window
     }
-
     func sceneDidEnterBackground(_ scene: UIScene) {
-      
         (UIApplication.shared.delegate as? AppDelegate)?.saveContext()
     }
-
-
 }
-
