+++ ToDoList/ToDoList/Models/TaskModel.swift
@@ -1,0 +1,8 @@
+//
+//  TaskModel.swift
+//  ToDoList
+//
+//  Created by Никита Данилович on 28.04.2023.
+//
+
+import Foundation
