+++ ToDoList/ToDoList/Resources/Fonts/DesignedSystemFonts.swift
@@ -1,0 +1,7 @@
+import UIKit
+
+struct DesignedSystemFonts {
+    static let headline: UIFont = UIFont.systemFont(ofSize: 24, weight: .bold)
+    static let body: UIFont = UIFont.systemFont(ofSize: 16, weight: .regular)
+    static let button: UIFont = UIFont.systemFont(ofSize: 18, weight: .bold)
+}
