--- ToDoList/ToDoList/Screens/Tasks/Presenter/TaskManagerViewController.xib
@@ -1,76 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21701" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_12" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21678"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="System colors in document resources" minToolsVersion="11.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="TaskManagerViewController" customModule="ToDoList" customModuleProvider="target">
-            <connections>
-                <outlet property="addTaskButton" destination="RJX-5G-4Xp" id="ndU-US-uSl"/>
-                <outlet property="taskTableView" destination="6kJ-8y-Qfn" id="6WE-Pa-obv"/>
-                <outlet property="view" destination="i5M-Pr-FkT" id="sfx-zR-JGt"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view clearsContextBeforeDrawing="NO" contentMode="scaleToFill" id="i5M-Pr-FkT">
-            <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="6kJ-8y-Qfn">
-                    <rect key="frame" x="0.0" y="59" width="393" height="759"/>
-                    <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                    <color key="sectionIndexBackgroundColor" systemColor="systemIndigoColor"/>
-                </tableView>
-                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="RJX-5G-4Xp">
-                    <rect key="frame" x="313" y="754" width="64" height="64"/>
-                    <constraints>
-                        <constraint firstAttribute="height" constant="64" id="Vgi-S0-0wu"/>
-                        <constraint firstAttribute="width" constant="64" id="cPH-MR-UBS"/>
-                    </constraints>
-                    <color key="tintColor" red="0.83137254901960778" green="0.30980392156862746" blue="0.30980392156862746" alpha="1" colorSpace="calibratedRGB"/>
-                    <inset key="imageEdgeInsets" minX="0.0" minY="0.0" maxX="2.2250738585072014e-308" maxY="0.0"/>
-                    <buttonConfiguration key="configuration" style="filled" cornerStyle="capsule">
-                        <color key="baseBackgroundColor" red="0.83137254900000002" green="0.30980392159999998" blue="0.30980392159999998" alpha="1" colorSpace="calibratedRGB"/>
-                    </buttonConfiguration>
-                    <connections>
-                        <action selector="addTaskTapped:" destination="-1" eventType="touchUpInside" id="16H-WV-z0v"/>
-                    </connections>
-                </button>
-                <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="plus" catalog="system" translatesAutoresizingMaskIntoConstraints="NO" id="5hj-zr-fA1">
-                    <rect key="frame" x="328" y="771" width="34" height="30"/>
-                    <color key="tintColor" red="1" green="1" blue="1" alpha="1" colorSpace="calibratedRGB"/>
-                </imageView>
-            </subviews>
-            <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
-            <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-            <color key="tintColor" red="0.83137254900000002" green="0.30980392159999998" blue="0.30980392159999998" alpha="1" colorSpace="calibratedRGB"/>
-            <constraints>
-                <constraint firstItem="5hj-zr-fA1" firstAttribute="leading" secondItem="RJX-5G-4Xp" secondAttribute="leading" constant="15" id="4Mr-ac-Mnh"/>
-                <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="6kJ-8y-Qfn" secondAttribute="trailing" id="BCr-52-amz"/>
-                <constraint firstItem="5hj-zr-fA1" firstAttribute="bottom" secondItem="RJX-5G-4Xp" secondAttribute="bottom" constant="-15" id="QUu-sf-xf1"/>
-                <constraint firstItem="6kJ-8y-Qfn" firstAttribute="leading" secondItem="fnl-2z-Ty3" secondAttribute="leading" id="SC9-hZ-bdK"/>
-                <constraint firstItem="fnl-2z-Ty3" firstAttribute="bottom" secondItem="6kJ-8y-Qfn" secondAttribute="bottom" id="Uya-Di-2AS"/>
-                <constraint firstItem="RJX-5G-4Xp" firstAttribute="bottom" secondItem="fnl-2z-Ty3" secondAttribute="bottom" id="WN1-ph-MdP"/>
-                <constraint firstItem="5hj-zr-fA1" firstAttribute="top" secondItem="RJX-5G-4Xp" secondAttribute="top" constant="15" id="ZK2-Wm-AnZ"/>
-                <constraint firstItem="5hj-zr-fA1" firstAttribute="trailing" secondItem="RJX-5G-4Xp" secondAttribute="trailing" constant="-15" id="pOX-Ys-3iJ"/>
-                <constraint firstItem="6kJ-8y-Qfn" firstAttribute="top" secondItem="fnl-2z-Ty3" secondAttribute="top" id="qFb-wF-ggp"/>
-                <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="RJX-5G-4Xp" secondAttribute="trailing" constant="16" id="qyR-C7-mK9"/>
-            </constraints>
-            <point key="canvasLocation" x="-16.030534351145036" y="19.718309859154932"/>
-        </view>
-    </objects>
-    <resources>
-        <image name="plus" catalog="system" width="128" height="113"/>
-        <systemColor name="systemBackgroundColor">
-            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-        </systemColor>
-        <systemColor name="systemIndigoColor">
-            <color red="0.34509803921568627" green="0.33725490196078434" blue="0.83921568627450982" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-        </systemColor>
-    </resources>
-</document>
