+++ ToDoList/ToDoList/Views/TaskTableViewCell/TaskTableViewCell.swift
@@ -1,17 +1,14 @@
 import UIKit
 
 class TaskTableViewCell: UITableViewCell {
-    
+
     static let identifier = "taskCell"
 
     override func awakeFromNib() {
         super.awakeFromNib()
-       
     }
 
     override func setSelected(_ selected: Bool, animated: Bool) {
         super.setSelected(selected, animated: animated)
-
     }
-    
 }
